# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APZ Legal Practice Management System API
security:
  - sessionCookie: []
  - bearerSession: []
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails/search:
    get:
      operationId: searchEmails
      tags: [email]
      summary: Search authorised Matter email
      parameters:
        - name: search
          in: query
          required: true
          schema: { type: string, minLength: 2 }
        - name: matterId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Matching email records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Email" }
...
  /emails/{emailId}/link:
    post:
      operationId: triageUnlinkedEmail
      tags: [email]
      summary: Link an unlinked email or deliberately leave it unlinked
      parameters:
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnlinkedEmailDecisionInput" }
      responses:
        "200":
          description: Triage decision recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Email" }
  /emails/ingest:
    post:
      operationId: ingestNormalizedEmail
      tags: [email]
      summary: Ingest a provider-normalized email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NormalizedEmailInput"
      responses:
        "201":
          description: Email ingested
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailIngestionResult" }
        "200":
          description: Existing provider message returned idempotently
        "409":
          description: No provider is connected
  /matters/{matterId}/emails:
    get:
      operationId: listMatterEmailsForMatter
      tags: [email]
      summary: List correspondence for an authorised Matter
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: unread
          in: query
          schema: { type: boolean }
        - name: linkStatus
          in: query
          schema: { type: string, enum: [unlinked, suggested, linked, reviewed_unlinked] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: Matter email records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Email" }
  /matters/{matterId}/emails/{emailId}:
    get:
      operationId: getMatterEmail
      tags: [email]
      summary: Get an authorised Matter email
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Email detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailDetail" }
  /matters/{matterId}/emails/{emailId}/read:
    patch:
      operationId: markMatterEmailReadForMatter
      tags: [email]
      summary: Mark an authorised Matter email read or unread
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isRead]
              properties:
                isRead: { type: boolean }
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Email" }
  /matters/{matterId}/emails/{emailId}/thread:
    get:
      operationId: getMatterEmailThread
      tags: [email]
      summary: Get the authorised Matter email thread
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Thread messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Email" }
  /matters/{matterId}/emails/{emailId}/link:
    post:
      operationId: linkMatterEmail
      tags: [email]
      summary: Link, move, or deliberately leave an email unlinked
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailLinkInput"
      responses:
        "200":
          description: Updated email link
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Email" }
  /email-attachments/{attachmentId}:
    get:
      operationId: getEmailAttachment
      tags: [email]
      summary: Securely retrieve an authorised email attachment
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Attachment bytes
        "403":
          description: Matter access denied
  /matters/{matterId}/emails/{emailId}/attachments/{attachmentId}/document:
    post:
      operationId: promoteEmailAttachmentToDocument
      tags: [email]
      summary: Create a governed Matter document from a stored email attachment
      parameters:
        - name: matterId
          in: path
          required: true
          schema: { type: integer }
        - name: emailId
          in: path
          required: true
          schema: { type: integer }
        - name: attachmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
      responses:
        "201":
          description: Governed document and traceability link
        "200":
          description: Existing governed document returned idempotently

  # ─── MICROSOFT 365 / OAUTH ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
    bearerSession:
      type: http
      scheme: bearer
      description: Opaque staff session token, not a JWT.
  schemas:
...
    Email:
      type: object
      required: [id, provider, externalMessageId, threadId, direction, senderEmail, receivedAt, linkStatus, isRead]
      properties:
        id: { type: integer }
        provider: { type: string }
        externalMessageId: { type: string }
        threadId: { type: integer }
        direction: { type: string, enum: [inbound, outbound] }
        senderEmail: { type: string }
        senderName: { type: ["string", "null"] }
        toRecipients:
          type: array
          items: { type: string }
        ccRecipients:
          type: array
          items: { type: string }
        bccRecipients:
          type: array
          items: { type: string }
        subject: { type: ["string", "null"] }
        bodyText: { type: string }
        bodyHtml: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"], format: date-time }
        receivedAt: { type: string, format: date-time }
        matterId: { type: ["integer", "null"] }
        clientId: { type: ["integer", "null"] }
        linkStatus: { type: string, enum: [unlinked, suggested, linked, reviewed_unlinked] }
        matchReason: { type: ["string", "null"] }
        matchConfidence: { type: ["integer", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EmailAttachment:
      type: object
      required: [id, emailId, filename, mimeType, fileSize, fileChecksum]
      properties:
        id: { type: integer }
        emailId: { type: integer }
        filename: { type: string }
        mimeType: { type: string }
        fileSize: { type: integer }
        fileChecksum: { type: string }
        documentId: { type: ["integer", "null"] }
        fileObjectPath: { type: ["string", "null"] }
    EmailLinkCandidate:
      type: object
      required: [id, emailId, matterId, score, reason, source, decision]
      properties:
        id: { type: integer }
        emailId: { type: integer }
        matterId: { type: integer }
        score: { type: integer }
        reason: { type: string }
        source: { type: string }
        decision: { type: string, enum: [pending, accepted, rejected] }
    EmailDetail:
      allOf:
        - $ref: "#/components/schemas/Email"
        - type: object
          required: [attachments, candidates]
          properties:
            attachments:
              type: array
              items: { $ref: "#/components/schemas/EmailAttachment" }
            candidates:
              type: array
              items: { $ref: "#/components/schemas/EmailLinkCandidate" }
...
    UnlinkedEmailDecisionInput:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [link, leave_unlinked] }
        targetMatterId: { type: integer }
    EmailLinkInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [link, change, remove, leave_unlinked]
        targetMatterId:
          type: integer
    NormalizedEmailInput:
      type: object
      required: [provider, externalMessageId, externalThreadId, senderEmail, receivedAt]
      properties:
        provider: { type: string }
        externalMessageId: { type: string }
        externalThreadId: { type: string }
        direction: { type: string, enum: [inbound, outbound] }
        senderEmail: { type: string }
        senderName: { type: string }
        toRecipients: { type: array, items: { type: string } }
        ccRecipients: { type: array, items: { type: string } }
        bccRecipients: { type: array, items: { type: string } }
        subject: { type: string }
        bodyText: { type: string }
        bodyHtml: { type: string }
        headers: { type: object, additionalProperties: { type: string } }
        sentAt: { type: string, format: date-time }
        receivedAt: { type: string, format: date-time }
        attachments:
          type: array
          items:
            type: object
            required: [filename, mimeType, fileSize, fileChecksum]
            properties:
              externalAttachmentId: { type: string }
              filename: { type: string }
              mimeType: { type: string }
              fileSize: { type: integer }
              fileChecksum: { type: string }
    EmailIngestionResult:
      type: object
      required: [email, duplicate]
      properties:
        email: { $ref: "#/components/schemas/Email" }
        duplicate: { type: boolean }
...